requirements.txt
opencv-python>=4.8.0
mediapipe>=0.10.0
numpy>=1.24.0
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
python-dotenv>=1.0.0
qdrant-client>=1.9.0
insightface>=0.7.3
onnxruntime>=1.16.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
aiofiles>=23.0.0
Jinja2>=3.1.0
